framework/requirements.txt
# Copyright (c) 2024 FORCERA, LDA
# 
# This program and the accompanying materials are made available under the
# terms of the Eclipse Public License 2.0 which is available at
# http://www.eclipse.org/legal/epl-2.0.
# 
# SPDX-License-Identifier: EPL-2.0

asyncpg==0.29.0
numpy==1.24.4
parrot-olympe==7.7.4
PySDL2==0.9.16
PyOpenGL
